move platforms into subdirectories
