+++ lib/platform/github/gh-got-wrapper.js
@@ -1,4 +1,4 @@
-const logger = require('../logger');
+const logger = require('../../logger');
 const ghGot = require('gh-got');
 const parseLinkHeader = require('parse-link-header');
