+++ lib/platform/github/index.js
@@ -1,4 +1,4 @@
-let logger = require('../logger');
+let logger = require('../../logger');
 const get = require('./gh-got-wrapper');
 
 const config = {};
