+++ lib/platform/gitlab/index.js
@@ -1,4 +1,4 @@
-let logger = require('../logger');
+let logger = require('../../logger');
 const get = require('./gl-got-wrapper');
 
 const config = {};
